Print the top left corner of a data frame

Prints the first n rows and columns of a data frame or matrix.

corner(x, n = 5)

Arguments

x
A data.frame.

n
The number of rows/columns to print.

Value

The corner of the data frame

Examples

corner(mtcars)

#>                    mpg cyl disp  hp drat
#> Mazda RX4         21.0   6  160 110 3.90
#> Mazda RX4 Wag     21.0   6  160 110 3.90
#> Datsun 710        22.8   4  108  93 3.85
#> Hornet 4 Drive    21.4   6  258 110 3.08
#> Hornet Sportabout 18.7   8  360 175 3.15

corner(iris, n=4)

#>   Sepal.Length Sepal.Width Petal.Length Petal.Width
#> 1          5.1         3.5          1.4         0.2
#> 2          4.9         3.0          1.4         0.2
#> 3          4.7         3.2          1.3         0.2
#> 4          4.6         3.1          1.5         0.2