Round numeric columns of a data frame
Source: R/dfsigfig.R

Round the numeric columns of a data frame to a specified number of significant digits.

dfsigfig(df, n = 3)

Arguments

df
A data.frame.

n
The number of significant digits to round off to.

Value

A data.frame rounded to n significant digits.

Examples

dfsigfig(mtcars,1)

#>                     mpg cyl disp  hp drat wt qsec vs am gear carb
#> Mazda RX4            20   6  200 100    4  3   20  0  1    4    4
#> Mazda RX4 Wag        20   6  200 100    4  3   20  0  1    4    4
#> Datsun 710           20   4  100  90    4  2   20  1  1    4    1
#> Hornet 4 Drive       20   6  300 100    3  3   20  1  0    3    1
#> Hornet Sportabout    20   8  400 200    3  3   20  0  0    3    2
#> Valiant              20   6  200 100    3  3   20  1  0    3    1
#> Duster 360           10   8  400 200    3  4   20  0  0    3    4
#> Merc 240D            20   4  100  60    4  3   20  1  0    4    2
#> Merc 230             20   4  100 100    4  3   20  1  0    4    2
#> Merc 280             20   6  200 100    4  3   20  1  0    4    4
#> Merc 280C            20   6  200 100    4  3   20  1  0    4    4
#> Merc 450SE           20   8  300 200    3  4   20  0  0    3    3
#> Merc 450SL           20   8  300 200    3  4   20  0  0    3    3
#> Merc 450SLC          20   8  300 200    3  4   20  0  0    3    3
#> Cadillac Fleetwood   10   8  500 200    3  5   20  0  0    3    4
#> Lincoln Continental  10   8  500 200    3  5   20  0  0    3    4
#> Chrysler Imperial    10   8  400 200    3  5   20  0  0    3    4
#> Fiat 128             30   4   80  70    4  2   20  1  1    4    1
#>  [ reached getOption("max.print") -- omitted 14 rows ]