Create tables in dokuwiki format

Prints the supplied data frame or matrix using Dokuwiki's table syntax, optionally copying the data to the clipboard (Mac OS X only).

dokuwiki(x, headersep = "^", sep = "|", clip = TRUE, ...)

Arguments

x
A data.frame.

headersep
The separator used between entries in the header row.

sep
The separator used between entries in all other rows.

clip
Whether or not to write the returned table to the clipboard (currently only supported on Mac OS X).

...
Further arguments passed to write.table.

Examples

dokuwiki(head(iris), clip=FALSE)

#> ^Sepal.Length^Sepal.Width^Petal.Length^Petal.Width^Species^
#> |5.1|3.5|1.4|0.2|setosa|
#> |4.9|3|1.4|0.2|setosa|
#> |4.7|3.2|1.3|0.2|setosa|
#> |4.6|3.1|1.5|0.2|setosa|
#> |5|3.6|1.4|0.2|setosa|
#> |5.4|3.9|1.7|0.4|setosa|

dokuwiki(head(mtcars), clip=FALSE, row.names=TRUE)

#> ^ ^mpg^cyl^disp^hp^drat^wt^qsec^vs^am^gear^carb^
#> |Mazda RX4|21|6|160|110|3.9|2.62|16.46|0|1|4|4|
#> |Mazda RX4 Wag|21|6|160|110|3.9|2.875|17.02|0|1|4|4|
#> |Datsun 710|22.8|4|108|93|3.85|2.32|18.61|1|1|4|1|
#> |Hornet 4 Drive|21.4|6|258|110|3.08|3.215|19.44|1|0|3|1|
#> |Hornet Sportabout|18.7|8|360|175|3.15|3.44|17.02|0|0|3|2|
#> |Valiant|18.1|6|225|105|2.76|3.46|20.22|1|0|3|1|