Improved list of objects

Improved list of objects. Sorts by size by default. This was shamelessly stolen from http://stackoverflow.com/q/1358003/654296.

lsa(pos = 1, pattern, order.by = "Size", decreasing = TRUE, head = TRUE,
  n = 10)

Arguments

pos
numeric. Position in the stack.

pattern
Regex to filter the objects by.

order.by
character. Either 'Type', 'Size', 'PrettySize', 'Rows', or 'Columns'. This will dictate how the output is ordered.

decreasing
logical. Should the output be displayed in decreasing order?

head
logical. Use head on the output?

n
numeric. Number of objects to display is head is TRUE.

Value

A data.frame with type, size in bytes, human-readable size, rows, and columns of every object in the environment.

Examples

not_run({
  a <- rnorm(100000)
  b <- matrix(1, 1000, 100)
  lsa()
})