Rownames to symbol-probeID

This function takes an exprs(eset) matrix where the rownames are probeset IDs and takes an annotated topTable output where you have an ID and Symbol column and outputs a character vector with symbol_probeid for each probeid in rownames(exprs(eset)). You can use this such that the output on a heatmap contains the gene names concatenated to the probe ID in case you have multiple symbols with the same probeID.

rownames_to_symprobe(exprset, tt)

Arguments

exprset
The output of exprs(eset).

tt
A topTable object.

Value

Character vector of the gene symbol with the probe ID.

Examples

# NOT RUN {
rownames_to_symprobe(esprs(eset), topTable(fit, number=nrow(fit)))
# }